Update text-fields.md
When adding a `TextInputLayout` with a floating label hint in landscape mode, the hint will only display correctly if a `TextInputEditText` is used instead of a regular `EditText`. [The official docs refer to "extract mode"](https://developer.android.com/reference/android/support/design/widget/TextInputEditText.html), which is when the EditText has focus in landscape mode and takes up the entire screen.